+++ versions.tf
@@ -1,8 +1,8 @@
 terraform {
-  required_version = ">= 0.13.1"
+  required_version = ">= 1.0.0"
 
   required_providers {
-    aws        = ">= 3.42.0"
+    aws        = ">= 3.45.0"
     local      = ">= 2.1.0"
     random     = ">= 3.1.0"
     kubernetes = "~> 2.2.0"